third_party/WebKit/LayoutTests/geolocation-api/cached-position-called-once-expected.txt

Tests that when a cached position is available the callback for getCurrentPosition is called only once. This is a regression test for http://crbug.com/311876 .

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isSuccess is true
PASS getCurrentPositionCallId is reportCount
PASS isSuccess is true
PASS getCurrentPositionCallId is reportCount
PASS isSuccess is true
PASS getCurrentPositionCallId is reportCount
PASS successfullyParsed is true

TEST COMPLETE
